requirements-runpod.txt
runpod
# Core HF stack for GLM-OCR (no vLLM)
torch==2.3.0
torchvision==0.18.0
transformers @ git+https://github.com/huggingface/transformers.git
accelerate
Pillow
requests
